+++ app/Http/Controllers/GerenteController.php
@@ -5,10 +5,11 @@ namespace App\Http\Controllers;
 use App\Models\Gerente;
 use App\Models\Pessoa;
 use App\Models\Endereco;
+use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Redirect;
-use Illuminate\View\View;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Contracts\View\View;
 
 class GerenteController extends Controller
 {
@@ -19,9 +20,10 @@ class GerenteController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Gerente::class);
         $gerentes = Gerente::all();
         return view('gerentes.index', ['gerentes' => $gerentes]);
-    }             
+    }
 
     /**
      * Show the form for creating a new resource.
@@ -30,6 +32,7 @@ class GerenteController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Gerente::class);
         return view('gerentes.create');
     }
 
@@ -37,26 +40,27 @@ class GerenteController extends Controller
      * Store a newly created resource in storage.
      *
      * @param Request $request
-     * @return \Illuminate\Http\Response
+     * @return RedirectResponse
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Gerente::class);
         $gerente = new Gerente();
         $pessoa = new Pessoa();
         $endereco = new Endereco();
-        
-        //validações 
+        $user = new User();
         $gerente->fill($request->validate(Gerente::$rules));
         $pessoa->fill($request->validate(Pessoa::$rules));
         $endereco->fill($request->validate(Endereco::$rules));
-
+        $request->merge(['tipo' => 'gerente']);
+        $user->fill($request->validate(User::$rules));
+        $user->password = Hash::make($user->password);
+        $user->save();
+        $pessoa->user()->associate($user);
         $pessoa->save();
-
         $gerente->pessoa()->associate($pessoa);
-
         $pessoa->endereco()->save($endereco);
         $gerente->save();
-
         return redirect()->action([GerenteController::class, 'show'], ['gerente' => $gerente]);
     }
 
@@ -68,6 +72,7 @@ class GerenteController extends Controller
      */
     public function show(Gerente $gerente)
     {
+        $this->authorize('view', $gerente);
         return view('gerentes.show', ['gerente' => $gerente]);
     }
 
@@ -79,6 +84,7 @@ class GerenteController extends Controller
      */
     public function edit(Gerente $gerente)
     {
+        $this->authorize('update', $gerente);
         return view('gerentes.edit', ['gerente' => $gerente]);
     }
 
@@ -87,10 +93,11 @@ class GerenteController extends Controller
      *
      * @param Request $request
      * @param Gerente $gerente
-     * @return View
+     * @return RedirectResponse
      */
     public function update(Request $request, Gerente $gerente)
     {
+        $this->authorize('update', $gerente);
         $gerente->fill($request->validate(Gerente::$rules));
         $gerente->save();
         $rules = Pessoa::$rules;
@@ -99,6 +106,8 @@ class GerenteController extends Controller
         $gerente->pessoa->save();
         $gerente->pessoa->endereco->fill($request->validate(Endereco::$rules));
         $gerente->pessoa->endereco->save();
+        $gerente->pessoa->user->fill($request->validate(User::$rules));
+        $gerente->pessoa->user->save();
         return redirect()->action([GerenteController::class, 'show'], ['gerente' => $gerente->refresh()]);
     }
 
@@ -110,6 +119,7 @@ class GerenteController extends Controller
      */
     public function destroy(Gerente $gerente)
     {
+        $this->authorize('delete', $gerente);
         $gerente->pessoa()->endereco()->delete();
         $gerente->delete();
         $gerente->pessoa()->delete();
